delete commit

+++ routes/taskRoutes.js
@@ -83,6 +83,10 @@ router.delete("/tasks/:id", authMiddleware, async (req, res) => {
             return res.status(404).json({ message: "Task not found" });
         }
 
+        if (task.userId.toString() !== req.user.id) {
+            return res.status(403).json({ message: "Not authorized to delete this task!" });
+        }
+
         await Task.findByIdAndDelete(id);
         res.json({ message: "Task deleted successfully" });
     } catch (error) {
